Group Mappers and DomDocuments in their own namespace and directory
