+++ src/ApiConnectors/SupplierApiConnector.php
@@ -4,8 +4,8 @@ namespace PhpTwinfield\ApiConnectors;
 
 use PhpTwinfield\Request as Request;
 use PhpTwinfield\Supplier;
-use PhpTwinfield\Supplier\SuppliersDocument;
-use PhpTwinfield\Supplier\SupplierMapper;
+use PhpTwinfield\DomDocuments\SuppliersDocument;
+use PhpTwinfield\Mappers\SupplierMapper;
 
 /**
  * A facade to make interaction with the the Twinfield service easier when trying to retrieve or send information about
