+++ src/Mappers/InvoiceMapper.php
@@ -1,11 +1,11 @@
 <?php
-namespace PhpTwinfield\Invoice;
+namespace PhpTwinfield\Mappers;
 
-use \PhpTwinfield\Invoice;
-use \PhpTwinfield\InvoiceLine;
-use \PhpTwinfield\InvoiceTotals;
-use \PhpTwinfield\Response\Response;
-use \PhpTwinfield\Customer;
+use PhpTwinfield\Invoice;
+use PhpTwinfield\InvoiceLine;
+use PhpTwinfield\InvoiceTotals;
+use PhpTwinfield\Response\Response;
+use PhpTwinfield\Customer;
 
 class InvoiceMapper
 {
